# docs/swagger.yaml
basePath: /api/v1
definitions:
  github_com_simonkefir_golang-messenger_internal_core_transport_http_response.ErrorResponse:
    properties:
      error:
        type: string
      message:
        type: string
    type: object
  internal_feature_chats_transport_http.ChatDTOResponse:
    properties:
      created_at:
        example: "2006-01-02 15:04:05"
        type: string
      id:
        example: 333
        type: integer
      participant:
        $ref: '#/definitions/internal_feature_chats_transport_http.ParticipantDTO'
    type: object
  internal_feature_chats_transport_http.ChatListItemDTO:
    properties:
      companion_id:
        example: 141
        type: integer
      companion_name:
        example: alex g
        type: string
      created_at:
        example: "2006-01-02 15:04:05"
        type: string
      id:
        example: 333
        type: integer
    type: object
  internal_feature_chats_transport_http.CreateChatDTO:
    properties:
      participant_id:
        example: 333
        type: integer
    required:
    - participant_id
    type: object
  internal_feature_chats_transport_http.ParticipantDTO:
    properties:
      display_name:
        example: alex g
        type: string
      user_id:
        example: 333
        type: integer
    type: object
  internal_feature_messages_transport_http.CreateMessageDTO:
    properties:
      content:
        example: mary is the girl, who i wanna f..
        maxLength: 4000
        minLength: 1
        type: string
    required:
    - content
    type: object
  internal_feature_messages_transport_http.MessageDTOResponse:
    properties:
      chat_id:
        example: 3
        type: integer
      content:
        example: its not what u r, its just what u did, dont hang up the phone, i
          lov3 u to death, eternal return...
        type: string
      id:
        example: 777
        type: integer
      sender_id:
        example: 333
        type: integer
      sent_at:
        example: "2006-01-02 15:04:05"
        type: string
    type: object
  internal_feature_messages_transport_http.PatchMessageDTO:
    properties:
      content:
        example: mary is the girl, who i wanna kiss
        maxLength: 4000
        minLength: 1
        type: string
      id:
        example: 17174
        minimum: 1
        type: integer
    required:
    - content
    - id
    type: object
...
host: 127.0.0.1:8080
info:
  contact: {}
  description: Messenger REST-API scheme
  title: Golang Messenger API
  version: "1.0"
paths:
  /chats:
    get:
      description: Получить информацию об о всех своих чатах из системы чатов
      produces:
      - application/json
      responses:
        "200":
          description: Успешно полученные чаты
          schema:
            $ref: '#/definitions/internal_feature_chats_transport_http.ChatListItemDTO'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_simonkefir_golang-messenger_internal_core_transport_http_response.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_simonkefir_golang-messenger_internal_core_transport_http_response.ErrorResponse'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/github_com_simonkefir_golang-messenger_internal_core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_simonkefir_golang-messenger_internal_core_transport_http_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Получить чаты
      tags:
      - chats
    post:
      consumes:
      - application/json
      description: Создать чат с участием другого пользователя
      parameters:
      - description: CreateChat тело запроса
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_feature_chats_transport_http.CreateChatDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Успешно созданный чат
          schema:
            $ref: '#/definitions/internal_feature_chats_transport_http.ChatDTOResponse'
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/github_com_simonkefir_golang-messenger_internal_core_transport_http_response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_simonkefir_golang-messenger_internal_core_transport_http_response.ErrorResponse'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/github_com_simonkefir_golang-messenger_internal_core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_simonkefir_golang-messenger_internal_core_transport_http_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Создать чат
      tags:
      - chats
  /chats/{chat_id}:
    delete:
      description: Удаление чата по его ID в системе чатов
      responses:
        "204":
          description: Успешно удалённый чат
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/github_com_simonkefir_golang-messenger_internal_core_transport_http_response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_simonkefir_golang-messenger_internal_core_transport_http_response.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_simonkefir_golang-messenger_internal_core_transport_http_response.ErrorResponse'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/github_com_simonkefir_golang-messenger_internal_core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_simonkefir_golang-messenger_internal_core_transport_http_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Удаление чата
      tags:
      - chats
    get:
      description: Получить информацию о своём чате из системы чатов
      produces:
      - application/json
      responses:
        "200":
          description: Успешно полученный чат
          schema:
            $ref: '#/definitions/internal_feature_chats_transport_http.ChatListItemDTO'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_simonkefir_golang-messenger_internal_core_transport_http_response.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_simonkefir_golang-messenger_internal_core_transport_http_response.ErrorResponse'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/github_com_simonkefir_golang-messenger_internal_core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_simonkefir_golang-messenger_internal_core_transport_http_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Получить чат
      tags:
      - chats
  /chats/{chat_id}/messages:
    delete:
      description: Удаление сообщения по ID чата и ID сообщения
      responses:
        "204":
          description: Успешное удаление сообщения в чате
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/github_com_simonkefir_golang-messenger_internal_core_transport_http_response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_simonkefir_golang-messenger_internal_core_transport_http_response.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_simonkefir_golang-messenger_internal_core_transport_http_response.ErrorResponse'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/github_com_simonkefir_golang-messenger_internal_core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_simonkefir_golang-messenger_internal_core_transport_http_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Удаление сообщения в чате
      tags:
      - messages
    get:
      description: Получить сообщения из чата по его ID
      parameters:
      - description: ID чата
        in: path
        name: chat_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Успешно полученные сообщения
          schema:
            items:
              $ref: '#/definitions/internal_feature_messages_transport_http.MessageDTOResponse'
            type: array
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/github_com_simonkefir_golang-messenger_internal_core_transport_http_response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_simonkefir_golang-messenger_internal_core_transport_http_response.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_simonkefir_golang-messenger_internal_core_transport_http_response.ErrorResponse'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/github_com_simonkefir_golang-messenger_internal_core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_simonkefir_golang-messenger_internal_core_transport_http_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Получить сообщения из чата
      tags:
      - messages
    patch:
      consumes:
      - application/json
      description: Изменить сообщение в чате по ID чата и ID сообщения
      parameters:
      - description: PatchMessage тело запроса
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_feature_messages_transport_http.PatchMessageDTO'
      produces:
      - application/json
      responses:
        "200":
          description: Успешно изменённое сообщение
          schema:
            $ref: '#/definitions/internal_feature_messages_transport_http.MessageDTOResponse'
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/github_com_simonkefir_golang-messenger_internal_core_transport_http_response.ErrorResponse'
        "401":
          description: unauthorized
          schema:
            $ref: '#/definitions/github_com_simonkefir_golang-messenger_internal_core_transport_http_response.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_simonkefir_golang-messenger_internal_core_transport_http_response.ErrorResponse'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/github_com_simonkefir_golang-messenger_internal_core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_simonkefir_golang-messenger_internal_core_transport_http_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Изменить сообщение
      tags:
      - messages
    post:
      consumes:
      - application/json
      description: Создать новое сообщение в чате по ID чата
      parameters:
      - description: ID чата
        in: path
        name: chat_id
        required: true
        type: integer
      - description: CreateMessage тело запроса
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_feature_messages_transport_http.CreateMessageDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Успешно созданное сообщение
          schema:
            $ref: '#/definitions/internal_feature_messages_transport_http.MessageDTOResponse'
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/github_com_simonkefir_golang-messenger_internal_core_transport_http_response.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_simonkefir_golang-messenger_internal_core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_simonkefir_golang-messenger_internal_core_transport_http_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Отправить сообщение
      tags:
      - messages
...
securityDefinitions:
  BearerAuth:
    description: Используйте "Bearer <ваш JWT-токен>"
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"